requirements.txt
Pillow
watchdog
requests
python-dotenv